feat: Update default LLM model to fara-7b and adjust agent provider and prompt.
