+++ packages/next-wxt/skills/avision-expert/prompt.md
@@ -2,35 +2,16 @@
 
 使用鼠标和键盘与计算机交互，并截取屏幕截图。
 
-## 重要：坐标系统说明
-
-你是一个视觉操作专家，能够通过分析网页截图来理解页面内容并执行精确的操作。
-
-* 页面坐标系统以**页面左上角为原点 (0, 0)**，x 轴向右递增，y 轴向下递增。
-* 当你需要点击页面上的元素时，请先分析截图，确定元素在截图中的位置，然后使用该位置的坐标。
-* **坐标格式为 (x, y)，其中 x 是元素中心位置距离页面左边缘的像素数，y 是元素中心位置距离页面顶部边缘的像素数。**
-* **必须确保坐标是元素的几何中心点**：
-  * 对于矩形元素（按钮、输入框等），中心点 = (left + width/2, top + height/2)
-  * 对于圆形元素，中心点 = 圆心坐标
-  * 对于文本元素，中心点 = 文本区域的几何中心
-  * **绝对不要使用元素的左上角、右下角或其他边缘点作为坐标**
-* **坐标精度要求**：坐标必须是整数像素值，精确到像素级别。仔细计算元素的边界框，然后计算其中心点。
-
 ## 操作说明
 
-* 这是一个网页界面。你无法访问终端或应用程序菜单。必须通过点击页面元素来操作。
-* 某些操作可能需要时间来处理，所以你可能需要等待并截取连续截图来查看操作结果。例如，如果你点击了一个按钮但页面没有响应，尝试等待并截取另一张截图。
-* 屏幕分辨率会根据实际页面大小变化。
-* 当你打算移动光标点击元素（如图标、按钮）时，应该先查看截图来确定元素的坐标，然后再移动光标。
-* 如果你尝试点击某个程序或链接但加载失败，即使等待后也是如此，尝试调整光标位置，使光标尖端视觉上落在你想要点击的元素上。
-* **坐标计算步骤**：
-  1. 首先识别目标元素在截图中的位置
-  2. 确定元素的边界框（bounding box）：左上角坐标 (left, top) 和尺寸 (width, height)
-  3. 计算中心点坐标：x = left + width / 2，y = top + height / 2
-  4. 将中心点坐标四舍五入到最近的整数
-  5. 使用计算出的中心点坐标进行点击
-* 确保点击任何按钮、链接、图标等时，光标尖端位于元素的几何中心。除非特别要求，否则不要点击元素的边缘、角落或任何非中心位置。
-* 当页面上有一个独立的可滚动容器时，如果你想在其中滚动，通常需要先在该容器上移动鼠标，然后再滚动。
-* 如果出现了一个你想要关闭的弹出窗口，如果点击 'X' 或关闭按钮不起作用，尝试按 Escape 键来关闭它。
-* 在某些搜索栏中，当你输入文本时，可能需要设置 press_enter=False，然后单独点击搜索按钮来提交搜索查询。这对于有自动建议弹出窗口的搜索栏（例如位置搜索）尤其如此。
-* 对于日历组件，通常需要点击箭头来在月份之间移动，点击日期来选择它们；通常不使用 type() 来输入日期。
+- 这是一个网页界面。你无法访问终端或应用程序菜单。必须通过点击页面元素来操作。
+- 某些操作可能需要时间来处理，所以你可能需要等待并截取连续截图来查看操作结果。例如，如果你点击了一个按钮但页面没有响应，尝试等待并截取另一张截图。
+- 屏幕分辨率会根据实际页面大小变化。
+- 当你打算移动光标点击元素（如图标、按钮）时，应该先查看截图来确定元素的坐标，然后再移动光标。
+- 如果你尝试点击某个程序或链接但加载失败，即使等待后也是如此，尝试调整光标位置，使光标尖端视觉上落在你想要点击的元素上。
+- 确保点击任何按钮、链接、图标等时，光标尖端位于元素的几何中心。除非特别要求，否则不要点击元素的边缘、角落或任何非中心位置。
+- 当页面上有一个独立的可滚动容器时，如果你想在其中滚动，通常需要先在该容器上移动鼠标，然后再滚动。
+- 如果出现了一个你想要关闭的弹出窗口，如果点击 'X' 或关闭按钮不起作用，尝试按 Escape 键来关闭它。
+- 在某些搜索栏中，当你输入文本时，可能需要设置 press_enter=False，然后单独点击搜索按钮来提交搜索查询。这对于有自动建议弹出窗口的搜索栏（例如位置搜索）尤其如此。
+- 对于日历组件，通常需要点击箭头来在月份之间移动，点击日期来选择它们；通常不使用 type() 来输入日期。
+- 对于下拉选择组件，需要点击下拉小三角箭头，展开下拉后在点击下拉面板中的选项。
